+++ CHANGELOG.md
@@ -10,6 +10,7 @@ All user visible changes to this project will be documented in this file.
  - Bumped dependencies (including nushell)
  - Added `expiry` option to `kv upsert`
  - Override `version` command to show cbsh version
+ - Add simple `query advise` subcommand
 
 ## 0.1.0 - 2020-05-20